agent-service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
pytest
google-generativeai
langchain
langchain-google-genai
pydantic-settings
